Releasing 2 New Scripts
PLEASE ADD BOTH APILOGIC AND LIBINLINE TO THE ONE CLICK

APILogic gives users command-line logic strutures (if, elseif, else), as well as term definition for text replacement and named condition groups for re-use later in the processing.

===APILogic REQUIRES libInline, and expects it to be named "libinline" in the approved.yaml (also updated in this push/pull).===

libInline provides a convenient interface for working with inlinerolls from the message object.